feat: switch to HandlerBoss instead of using InitialHandler.
