+++ sonar-bungee/src/main/java/xyz/jonesdev/sonar/bungee/fallback/FallbackUserWrapper.java
@@ -22,6 +22,7 @@ import io.netty.channel.ChannelPipeline;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import net.kyori.adventure.text.Component;
+import net.md_5.bungee.connection.InitialHandler;
 import org.jetbrains.annotations.NotNull;
 import xyz.jonesdev.sonar.api.fallback.Fallback;
 import xyz.jonesdev.sonar.api.fallback.FallbackUser;
@@ -32,16 +33,17 @@ import java.net.InetAddress;
 
 @Getter
 @RequiredArgsConstructor
-public final class FallbackUserWrapper implements FallbackUser<FallbackInitialHandler> {
+public final class FallbackUserWrapper implements FallbackUser<InitialHandler> {
   private final Fallback fallback;
-  private final FallbackInitialHandler connection;
+  private final InitialHandler connection;
   private final Channel channel;
   private final ChannelPipeline pipeline;
   private final InetAddress inetAddress;
   private final ProtocolVersion protocolVersion;
+  private final FallbackHandlerBoss handlerBoss;
 
   @Override
   public void disconnect(final @NotNull Component reason) {
-    connection.closeWith(Disconnect.create(reason));
+    handlerBoss.closeWith(Disconnect.create(reason));
   }
 }
